Fix SUB_Remove deferred thinks
Fix #330

Previously, we were overriding `SUB_Remove` for NT guns as a hack to
prevent them disappearing mid-round. This fix re-enables the correct
`SUB_Remove` behaviour, and instead handles the deferred removals at the
`SetPickupTouch` override, based on `mp_weaponstay` cvar's boolean
value - with the exception of the ghost never being deleted from the
world, regardless of the cvar's value.

We're defaulting to the more sensible `mp_weaponstay 1` behaviour to
keep dropped guns in the world over parity, but the original NT
behaviour can be restored by toggling this cvar to 0 instead.