Workaround for Linux shaders breakage
Fix #438

This is a hack around the vision mode shaders sporadically breaking on
Linux builds.

It seems that for whatever reason, the shader state cached at the
glshaders file somehow leads to the corruption upon successive runs of
the game, and write-protecting the file offers a simple way to avoid
that breakage.

Ideally, we should figure out  what's really going wrong here at the
shader system and/or precaching level, but this should suffice for
avoiding the bug.

We're doing this at runtime because we can't version control the
filesystem specific file metadata required for the read-only state.

For more discussion, see:
* https://github.com/NeotokyoRebuild/neo/pull/587
* https://github.com/NeotokyoRebuild/neo/issues/438#issuecomment-2351068151