Default-disable glshaders.cfg generation
Turn off the cvar for glshaders.cfg file generation on exit by default.

Also remove the file, if it exists on launch. This will be too late to
take effect for runs of the game where the file already exists, but
should fix it for future runs.

This also removes the dependency on filesystem specific read permission
flags for the fix.