Fix grenade bugs
Fix #708
Fix #713

Fix problems with throwable items like grenades.

This PR disables priming throwables with the +attack2/"alt fire" button, as that seemed to be happening by mistake.

This also removes the experimental underhanding/lobbing of grenades feature. This was never really properly implemented, and was just remnants from tests years ago. It's not parity, and if we wanted this kind of system in the future, it should probably go through a proper design doc process to be added.

Finally, this fixes the inability to throw grenades/throwables whilst sprinting - parity behaviour is to allow throwing of nades regardless of sprint state.