Fix knife incorrect initial secondary attack
Fix a problem where the knife swing could be incorrectly triggered once
using the `+attack2` input.

This is a supplementary fix to #708 since I forgot to address the knife
stuff in the earlier PR #714 that closed it.

We had some buggy scaffolding in place for a speculative future
secondary attack for the knife, but in the spirit of YAGNI I've decided
to delete it for now to simplify the code.