Fix main menu maxplayers slider value
Since the x64 update 2fdf6fb2, the engine's MAX_PLAYERS limit (for x64)
has increased from 33 to 101. Update the main menu's server creation
panel slider value to use this definition, instead of hardcoding 32 as
the upper bound.

We're using MAX_PLAYERS-1 because as per
https://github.com/NeotokyoRebuild/neo/blob/8f0fee0b0e4760af3851afb856815de9db2a9250/src/game/shared/shareddefs.h#L285-L287
the maximum gamerules allowed players should be <MAX_PLAYERS to make
room for the SourceTV/replay bot on the last allocated slot.