Revert snd_async_fullyasync to 0
Remove the autoexec that sets cvar snd_async_fullyasync to 1.

The default value for this cvar is 0. It was flipped on as a temp workaround
for some sound and/or disk IO related stutters when I was working on the
mod a long time ago with a really ancient PC.

The engine sound system is closed source and a bit of a black box, but
this change could be responsible for bug #1144

I'm seeing no negative effects in terms of gameplay smoothness, and
since this is the engine default value, we should most likely prefer to
use it as our default anyway.