Fix late-binding UI elements
The Neo UI elements were only bound once, but if for example the MP3
menu didn't have a bind when that bound logic ran, it would never get
associated with a key.

Instead, refresh the binds every time, so that they're always guaranteed
to be up to date, even if the user re-binds the key after it had been
bound (or was previously unbound).

We still have the problem here that the user can't assign multiple
keybinds for the same command, because the latest bind will always take
precedence.

Possibly(?) related: https://github.com/NeotokyoRebuild/neo/issues/1166