Fix UI bind bugs (#1233)
* Fix unbound keys triggering unbound UI elements

If the user inputs a key that has no mapping in the engine, that key
will result in BUTTON_CODE_NONE. Any UI element which the user has not
associated with a key will also map to BUTTON_CODE_NONE. Therefore,
pressing these kinds of keys (such as multimedia keys like "Volume Up")
can incorrectly trigger UI elements that are un-bound, for example the
mp3 player UI.

This change will early return for any inputs where the key code inputted
was interpreted as BUTTON_CODE_NONE.

* Fix late-binding UI elements

The Neo UI elements were only bound once, but if for example the MP3
menu didn't have a bind when that bound logic ran, it would never get
associated with a key.

Instead, refresh the binds every time, so that they're always guaranteed
to be up to date, even if the user re-binds the key after it had been
bound (or was previously unbound).

We still have the problem here that the user can't assign multiple
keybinds for the same command, because the latest bind will always take
precedence.

Possibly(?) related: https://github.com/NeotokyoRebuild/neo/issues/1166

* Add comment

* refactor