Prevent dead-owned guns from rendering in world
We had a bug where the dead spectators could somehow end up possessing
weapons, and these weapons would then appear floating around the world
wherever the spectators flew. This is just a failsafe to stop rendering
such guns, should that ever happen.

We also have a ::BumpWeapon safety check in place now, so hopefully the
spectators should no longer be able to end up with a weapon in their
inventory in the first place, but you never know...