Display takeover player name in scoreboard/feed
When a spectator takes over another player, display their name with the format of: "target_name (spectator_name)"

Swap stats on the scoreboard when taking over another player

This has the effect of making it seem like a spectator really took the place of a bot/player, while their original scoreboard place looks like them but deceased.

Helper for displaying full takeover context names

If a player is taken over by a spectator, display the full context of the takeover relationship in the killfeed, whether the taken over player is a killer, victim, or goal captor.