Fix duplicate weps for non-CTG modes (#1397)
Fix regression from 20b5821

We need to run the BaseClass code for CNEO_Player::BumpWeapon
even in the IsDead case, because the current weapon spawning logic is
reliant on it for game modes other than CTG.

This is just a hotfix, this whole thing could use a rework. For example
it appears we apply GiveDefaultItems several times pointlessly, which
could be cleaned up also.