Reset explosion SFX client-side on spawn
Reset the ear-ringing sound effect caused by standing near
grenade/det explosions at spawn/respawn at client side. The server
already does this by default via `CBasePlayer::Spawn`, but we want to
ensure that reset arrives on time (and reliably; the default recipient
filter is not marked reliable) for the remote client.

This is done to address #1376

The reason this bug seems difficult to reproduce may be related to the
fact the server-sided recipient filter wasn't marked reliable, so you
might need a dropped packet for it to occur, but this is just speculation on my part.

For alternatives considered, see the orphaned commits from before the
rebase, and the dicussion at #1383:
* 0d4ef5f13e9aba8df7b96722782f199e018371ff
* 3f3f32f99dac7d2726fc09bace7f7101e841f87a
* 564a57c61b76effe53e5efdc3918c174c44d4940