Add commands for adjusting bots' class (#1478)
* Add commands for adjusting bots' class

* Add `bot_class` cvar for adjusting which class bots should spawn as the
next time. This var name and behaviour mimics the `bot_class` cvar of
OGNT, with the exception of starting with value of -1 (don't change,
instead use random) instead of 0 (always spawn recon bots).

* Add `bot_changeclass` command for switching all bots' classes right
  now instead of at next spawn (if allowed by gamerules/other cvars).

* Only request team change for bots in game team

* improve asserts

* Fix random class logic

Make sure the randomized spawns are respected when either the ConCommand
is used, or the ConVar is returned to the -1 (random team) value.

* Fix class limits to exclude Juggernaut spawns

* cleanup

* Set bot class at Spawn instead of CNEOBotManager

Because we have to we able to adjust bot spawn class controlled by
cvars, I'm removing the responsibility from CNEOBotManager to choose the
bot class (used to fire only once at bot creation), and instead check it
separately for each new spawn of that bot.

This has the (positive?) side-effect that bots will reconsider which
class they should be playing, taking the class-specific ratio cvars into
consideration.