Add "nwae" option
If the branch name or the latest commit message contains the phrase "nwae"
(short for: no-warnings-as-errors), allow CI to skip warnings-as-errors for the build.
This is inteded as means for devs to temporarily dodge the warning rules in exceptional
situations (breakage from compiler version changes etc), but the preferred solution almost
always is to actually fix the warning instead.