Make unmanned JGR solid, with soft-collide post-death transitions
This commit makes the JGR prop solid when it is uncontrolled and idle in
the world. Once a player occupies the JGR, "it" starts abiding by the
usual player collision rules. Once the JGR dies, it will transition back
to being solid through a brief period of soft-collision, which is used
to avoid nearby players getting stuck inside the collider during the
transition.

The soft collisions are implemented as a child point_push object, which
will push away any players inside the collider until the JGR think
routine has confirmed its vicinity is clear for becoming solid again.