Fix neo_restart_this leaking edicts
For some reason, calling RestartGame directly from the neo_restart_this
change callback does not clean up edicts correctly, which leads to an
edict starvation and a subsequent game crash when using "neo_restart_this
1" repeatedly in a sufficiently complex map.

Because we don't have support for the SDK native
"mp_restartgame_immediate", and we don't want "neo_restart_this" to
mimic the SDK native "mp_restartgame" center of screen printout, this
commit works around those limitations by calling "mp_restartgame" with a
magic value that we intercept from CHL2MPRules::CheckRestartGame, and
then monkey with the values to achieve parity-ish map reset behaviour
while sidestepping the edict leak problem.

This is a complete hack and should be revisited, but since the command
is essentially broken currently, this is better than nothing for the
time being.